fix(dataframe_utils): roll_df() return None instead of the dataframe_utils
In case when shift parameter is 0 roll_df() method may not to do any
modifications with given dataframe and just return it back. But due to
empty return in code in this case appears the bug when insted of the given
dataframe the method returns None. The commit fix it, so now the method
works correct.