feat(data worker): add get_batch_generator() to convert generator to batch one
For example the get_crop_generator() return crops i.e. np.ndarrays of shape
(rows,cols,channels) but it is not so efficient and fast. In the other hand
due to data quantity there are cases when we can't nether use generator as it
is slow or numpy array as it uses to much RAM and PC can't handle it so
the get_batch_generator() was created to overcomming the problem. It can
divide generator data flow to baches of size that the PC can store in RAM,
and due to bach storing in RAM it computations is way faster than raw
genetator case