feat(custom_modules.dataset): delete get_augmented_crop_generator(), get_batch_generator
The get_augmented_crop_generator() appears to work uncorrect. Due
to all used augmentations half of the data having been created by
it was dublicates. The one could use get_crop_generator() with
the augmentatios=True param to get the same functionality but without
the bug described earier.

The get_batch_generator() func looks redundunt due to existing
the batch dividing functionality in tensorflow and pytorch datasets.

The model_train_with_generator.ipynb used the deleted functions and
was codded oddly. The file functionality will be rewritten if it
nessary with new functions